requirements.txt
flask==2.3.3
flask-cors==4.0.0
flask-cors==4.0.0
google-generativeai==0.8.5 
firebase-admin==6.5.0
firebase-admin==6.5.0
requests==2.31.0
yt-dlp==2025.12.08
pillow
yfinance==0.2.18
python-dotenv==1.0.0